bsr(api): migrate Domain errors in shared into shared/error-manager
