+++ api/src/shared/application/http-errors.js
@@ -7,7 +7,6 @@ class BaseHttpError extends Error {
     super(message);
     this.title = 'Default Bad Request';
     this.status = 400;
-    this.fromShared = true; // TODO: delete this property after migration to shared complete
   }
 }
