+++ api/tests/shared/unit/application/error-manager_test.js
@@ -4,6 +4,7 @@ import { NotFoundError } from '../../../../src/shared/domain/errors.js';
 
 import { HttpErrors } from '../../../../src/shared/application/http-errors.js';
 import { handle } from '../../../../src/shared/application/error-manager.js';
+
 describe('Shared | Unit | Application | ErrorManager', function () {
   describe('#_mapToHttpError', function () {
     it('should instantiate NotFoundError when NotFoundError', async function () {
